+++ sdks/go/node/containerruntime/k8s/k8s.go
@@ -41,8 +41,11 @@ type _containerRuntime struct {
 func (cr _containerRuntime) Delete(
 	ctx context.Context,
 ) error {
-	// for now this is a no-op
-	return nil
+	return cr.k8sClient.CoreV1().Pods("opctl").DeleteCollection(
+		ctx,
+		*metaV1.NewDeleteOptions(0),
+		metaV1.ListOptions{},
+	)
 }
 
 func (cr _containerRuntime) DeleteContainerIfExists(
@@ -52,7 +55,7 @@ func (cr _containerRuntime) DeleteContainerIfExists(
 	if err := cr.k8sClient.CoreV1().Pods("opctl").Delete(
 		ctx,
 		constructPodName(containerID),
-		metaV1.DeleteOptions{},
+		*metaV1.NewDeleteOptions(0),
 	); err != nil {
 		return fmt.Errorf("unable to delete k8s container: %w", err)
 	}
@@ -63,8 +66,7 @@ func (cr _containerRuntime) DeleteContainerIfExists(
 func (cr _containerRuntime) Kill(
 	ctx context.Context,
 ) error {
-	// for now this is a no-op
-	return nil
+	return cr.Delete(ctx)
 }
 
 func (cr _containerRuntime) RunContainer(
@@ -82,6 +84,15 @@ func (cr _containerRuntime) RunContainer(
 	if err != nil {
 		return nil, err
 	}
+
+	// don't leak terminated pods
+	defer cr.k8sClient.CoreV1().Pods("opctl").DeleteCollection(
+		ctx,
+		*metaV1.NewDeleteOptions(0),
+		metaV1.ListOptions{
+			FieldSelector: "status.phase!=" + string(coreV1.PodPending) + ",status.phase!=" + string(coreV1.PodRunning) + ",status.phase!=" + string(coreV1.PodUnknown),
+		},
+	)
 
 	_, err = cr.k8sClient.CoreV1().Pods("opctl").Create(
 		ctx,